+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.11.33
+
+* bugfix release: a fix for a certain bug involving `get_arguments` and `classmethod`